pkg/catalog/directory_loader.go: skip hidden nodes
Hidden directories and hidden files (as in, nodes starting with the
character `.`) should be skipped when loading in CSVs and CRDs. This
allows things like a Kubernetes volume mount of a config map to work,
because otherwise the CSV would be discovered twice (the real file, and
the symlink to it) and fail CRD ownership validation the second time.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 8e5159833c65b2a829b7cf9ca83b569f6e8f5e44