fix(resolver): resolved dependent subscriptions don't detect updates
this occurred because we only check for updates when we have information
about the currently installed CSV on the subscription status, which
was omitted from dependent catalogs.

the resolver now records the csv resolved into the subscription spec,
and it is used to populate the installed csv status when available.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 4b339b8946c4d7d80fab08b56f5fb987df51d256