fix(unit): operator group test flakes
This forces TestSyncOperatorGroups to allow the cache to catch up as the
number of resources that are checked here is high.

The additional ignoreCopyError parameter in the tests have been added to
handle the scenario where syncCopyCSV would never be executed in a
production workflow, but for testing it's being called explicitly.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 84c8914e3166d888d8e6c7225cd7e2858817a784