fix(olm): clean up resources on CSV deletion
This ensures proper resource deletion is done upon CSV deletion. Since
this touches a lot of different places, here's a summary of changes
made:

The RBAC has been modified to be owned by CSV instead of the operator
group. An operator group may remain after a CSV is deleted, but the
associated resources shouldn't. Similarly, created service accounts were
missing an owner reference to the CSV.

Due to the large amount of CSV requeueing and potential in progress
handling of a CSV, RBAC couldn't be deleted in
handleClusterServiceVersionDeletion (because sometimes the RBAC would be
recreated by another CSV sync). Instead, a new queue was created for
GC-ing resources. The sync loop specifically is used to do deletes so
that the loop can return an error (an error being if the CSV is not yet
deleted) and will be scheduled to try again later. The requeueing code
has been changed to not requeue if the CSV is not in the cache to help
not delay the new GC sync loop.

The new queue does not utilize an informer or indexer, so the event and
the resource are placed directly on the queue rather than relying on the
indexer to retrieve by key in the processing loop (processNextWorkItem).

Upstream-repository: operator-lifecycle-manager
Upstream-commit: f182d0e44dfeb01b91bc70256fe602196143b625