test(e2e): add operator group SA permission tests
The first test ensures that a service account with insufficient
permissions is properly handled after necessary RBAC is added to the
service account.

The second test ensures that an operator group with a service account
that is lacking permissions then properly installs a CSV after the limiting
service account is removed from the spec.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 858e619ef98f7cd3ee8a2557abb4e7f768de531a