feat(subscription) Config Volume and VolumeMounts
This commit updates the SubscriptionConfig type to include the Volumes
and VolumeMounts fields. These fields can be used to set the Volume and
VolumeMount fields on operators deployed by OLM.

This commit also introduces a number of unit tests and an e2e test that
ensures that this feature is working as intended.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 07253705ff05828c930b2e4f6bdb1c68c0649bdf