Fix kind detection for local builds.
By default, kind 0.7.0 context names are of the form
kind-* (e.g. kind-kind). Not sure when this changed -- presumably the
behavior here worked at the time it was written.

This also skips minikube setup if the current context is guessed to
point to a kind cluster, without requiring nonempty NO_MINIKUBE.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: b6958b3565ddd4e9c8ecea4fb7be624e37d54843