Update chart to use the new "crds" directory.
Helm 3 will handle CRD installation for CRD manifests located in the
chart's "crds" subdirectory, in exchange for dropping template
rendering support for CRD manifests. The only CRD template that exists
in the OLM chart today is used by a single test of downstream-only
behavior. The others are moved to "crds", and the release script has
been modified to look in both "templates" and "crds" for
manifests. The apiVersion in Chart.yaml has accordingly been bumped to
v2 to indicate that it requires Helm 3.

On its own, this change is intended to have minimal impact, but it's
part of the groundwork for bootstrapping E2E test environments for our
upstream CI.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 1bf7d627df4c9a21fbd8555a4fe7fa6d7d0df9e8