fix(resolver): don't pick operator from same package when resolve requiredAPI
When resolver looks for an operator that provides requiredAPIs, it
doesn't exclude operators from the same package that the original
operator that requires those APIs. As a result, if there is an
existing version of that operator now provides those requiredAPIs
(due to API ownership transfer), resolver will select an older or
newer version of the same operator to fulfill that requiredAPIs
which leads to API ownership conflicts on providedAPIs.

The resolver will look up all latest channel entries that provides
the necessary requiredAPIs and filter out those from the same package
before go on with the operator selection.

Signed-off-by: Vu Dinh <vdinh@redhat.com>

Upstream-repository: operator-lifecycle-manager
Upstream-commit: e899b6504b075c9a1fe3ceda325f4ae4bd19faa9