Bug 1847540: resolve only default channels
Non deterministic behavior was introduced where, in some cases, the
dependency resolver would select a channel entry from the registry
that was not part of the default channel and create a subscription
on that channel. This bugfix updates the registry client's filter
function to ignore non default channel entries.

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 44df9674df798f17d19330a28048539edb4349f9