Add more precise matching to the metric end-to-end tests.
The metric tests scrape metrics in the Prometheus text format and have
been mainly asserting based on the presence or absence of substrings
within the entire response body.

Metrics are now parsed into a slice of structs, and there is a new
Gomega matcher that can reference metric family, labels, and
value (for untyped metrics, gauges, and counters).

Upstream-repository: operator-lifecycle-manager
Upstream-commit: d8527b2efe7f85e10166006437ec4332727c66f8