Update index add func to handle optional default channel
If default channel is not provided in annotations, the existing
default channel will remain the same.

For the case of the first bundle being added and the default channel
isn't provided, it will be inferred from provided channel list if
there is only one channel. Otherwise, the index add will fail until
the default channel is provided.

Signed-off-by: Vu Dinh <vdinh@redhat.com>

Upstream-repository: operator-registry
Upstream-commit: 4d61b1aab171c890849090f7f7398586200bcbd6