Give more precise reasons for Subscriptions with no candidates.
A Subscription specifies certain criteria for selecting operators from
catalogs. The granularity of the Subscription criteria is, from least
to most specific: catalog, package, channel, startingCSV (if
applicable). If a user creates a Subscription and its criteria does
not match a single operator, the failure message now indicates the
highest level of granularity that excludes all available
operators. For example, if a Subscription specifies a package name
that doesn't exist in the referenced, catalog, the specified package
name will appear in the failure text.

Signed-off-by: Ben Luddy <bluddy@redhat.com>

Upstream-repository: operator-lifecycle-manager
Upstream-commit: a486e5435e6a9ae281859c83ef5d85a2a28d96ab