Fail InstallPlan on bundle unpack timeout
The InstallPlan sync can stay stalled on the Installing phase if the bundle cannot be successfully unpacked.
Adding a configurable timeout for the duration of the bundle unpack Job helps identify if an unpack Job is
stalled, and the InstallPlan is then transitioned to Failed with the unpack Job's failure condition
propagated to the InstallPlan condition.

Upstream-commit: 27ced4137bc4c7637b9c36f95946fe53cefe0e3d
Upstream-repository: operator-lifecycle-manager