Delete the symbolically linked operator-registry.Dockerfile
This was a temporary workaround while we split up the upstream/downstream CI promotion. Previously, we had specified the `Dockerfile` as the name for the registry CI/ART images, but when we started moving everything to the downstream monorepo, we had multiple Dockerfile(s) (e.g. operator-lifecycle-manager.Dockerfile, operator-registry.Dockerfile) and the auto-config-brancher bot was ensuring Dockerfile(s) matching that list of names was present in this repository, which they were not.

Upstream-repository: operator-registry
Upstream-commit: b38c6f4c0df4311df439b2fbb6286c658a3d5cf9