scripts: Update the scripts/generate_crds_manifests.sh
Cleanup the bash script responsible for rendering the list of helm charts
that are contained in the staging/operator-lifecycle-manager path.

Update that script to also make several inplace changes to the
PackageServer CSV YAML manifest after it's been rendered from the
staging path. This is to avoid making any changes to the corresponding
helm chart template for the CSV resource in the staging path, and to
have the verify-manifests target pass locally and in CI.