Revert everything back to the initial commit
This is necessary to get the repository back into a state where all
commits introduced after this commit have the requisite metadata in
order to identify which commits, and in what order, are present in
downstream when porting upstream commit(s).

This is a stop gap in order to avoid force pushing a "clean" commit
history entirely. That approach had gotten rejected due to concerns with
it inadvertently affecting the downstream build process.