go.*,vendor: Update the o/api, k8s, and controller-runtime dependencies
Update the o/api dependency to a newer version that includes the control
plane topology updates. We need to add a replace directive as o/api
doesn't follow proper semantic version tagging. That replace directive
points to a commit in the release-4.8 branch.

Update the k8s APIs to a v1.22 beta release.

Update controller-runtime to v0.9.x as it's incompatible with API
changes made in the k8s APIs.