Add managed label to configmap to reduce cache size
Using managed label to filter relevant configmaps in the cluster
to reduce the configmap cache size.

For catalogsource configmaps, they are created by users and there are
no prexisting labels, there are extra API calls initially to add
the label to migrate those configmaps to the catalog cache.

Signed-off-by: Vu Dinh <vudinh@outlook.com>

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 9d7732c6b711530aaa051a8aba34af8128d2cf14