Prune unused ClusterServiceVersion fields from catalog's informer.
The catalog operator watches both copied and original
ClusterServiceVersions in order to perform dependency resolution and
to check for API ownership conflicts during installation, but it only
reads a small subset of fields. Pruning those fields as CSVs are read
from the wire can substantially reduce heap size on clusters with a
significant total number of CSVs across all namespaces (for example,
when operators target all namespaces on a cluster with thousands of
namespaces).

Signed-off-by: Ben Luddy <bluddy@redhat.com>

Upstream-repository: operator-lifecycle-manager
Upstream-commit: 895e7ac738d081a73f107af9843b02cc7a6e0c2f